Fix empty staging argument which confuses certbot-auto
If you were not using the staging environment an empty string was used
as a command line argument for certbot-auto, which resulted in the
command failing. To fix this we now append the staging argument to the
arguments array so no argument is added on invocation in a non-staging
environment.