Fixed issue #19 (#20)
* Update renewAndSendToProxy.sh

Add a retry option in case curl fails

* Update renewAndSendToProxy.sh

* Update renewAndSendToProxy.sh

Fix updating the wrong variable

* Update renewAndSendToProxy.sh

Move counter update to the end of the loop, so our counter logic is
valid

* Update renewAndSendToProxy.sh

Add options to curl to hide the progress bar but show errors which can
be helpful

* Update renewAndSendToProxy.sh

Move counter back to the beginning as this avoids any logic in the print
statement

* Update renewAndSendToProxy.sh

Move sleep statement to avoid unnecessary sleep

* Update certbot.sh

- Add a dry-run (with a retry mechanism) to make sure Let's Encrypt is
able to validate the domain, as in some environments this might take a
while.
- To avoid specifying the common arguments to certbot multiple times store them
once.

* Fix empty staging argument which confuses certbot-auto

If you were not using the staging environment an empty string was used
as a command line argument for certbot-auto, which resulted in the
command failing. To fix this we now append the staging argument to the
arguments array so no argument is added on invocation in a non-staging
environment.